+++ dwds/lib/src/loaders/require.dart
@@ -6,6 +6,7 @@ import 'dart:convert';
 
 import 'package:dwds/src/debugging/dart_runtime_debugger.dart';
 import 'package:dwds/src/debugging/metadata/provider.dart';
+import 'package:dwds/src/loaders/asset_scheme.dart';
 import 'package:dwds/src/loaders/strategy.dart';
 import 'package:dwds/src/readers/asset_reader.dart';
 import 'package:dwds/src/services/expression_compiler.dart';
@@ -138,8 +139,10 @@ class RequireStrategy extends LoadStrategy {
     this._serverPathForAppUri,
     this._moduleInfoForProvider,
     AssetReader assetReader,
-    this._buildSettings, {
+    this._buildSettings,
+    this.assetScheme, {
     super.packageConfigPath,
+    super.metadataLoader,
   }) : super(assetReader);
 
   @override
@@ -161,6 +164,9 @@ class RequireStrategy extends LoadStrategy {
 
   @override
   String get moduleFormat => 'amd';
+
+  @override
+  final AssetScheme assetScheme;
 
   @override
   String get loadLibrariesModule => 'require.js';