+++ dwds/lib/src/readers/asset_reader.dart
@@ -2,12 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:collection/collection.dart';
 import 'package:dwds/src/utilities/shared.dart';
 import 'package:dwds/src/utilities/web_path_translator.dart';
 import 'package:file/file.dart';
 import 'package:logging/logging.dart';
 import 'package:package_config/package_config.dart';
+import 'package:path/path.dart' as p;
 export 'package:dwds/src/utilities/shared.dart' show stripLeadingSlashes;
 
 /// A reader for Dart sources and related source maps.
@@ -40,12 +40,8 @@ abstract class AssetReader {
   Future<void> close();
 }
 
-class PackageUriMapper {
-  final _logger = Logger('PackageUriMapper');
-  final PackageConfig packageConfig;
-  final bool useDebuggerModuleNames;
-
-  static Future<PackageUriMapper> create(
+abstract class PathResolver {
+  static Future<PathResolver> create(
     FileSystem fileSystem,
     Uri packageConfigFile, {
     bool useDebuggerModuleNames = false,
@@ -53,68 +49,266 @@ class PackageUriMapper {
     final packageConfig = await loadPackageConfig(
       fileSystem.file(packageConfigFile),
     );
-    return PackageUriMapper(
-      packageConfig,
+    return BuildRunnerPathResolver(
+      packageConfig: packageConfig,
       useDebuggerModuleNames: useDebuggerModuleNames,
     );
   }
 
-  PackageUriMapper(this.packageConfig, {this.useDebuggerModuleNames = false});
+  /// Computes the server path for a given application URL.
+  String? appUriToServerPath(String appUrl, {bool? useDebuggerModuleNames});
 
-  /// Compute server path for package uri.
-  ///
-  /// Note: needs to match `urlForComponentUri` in javascript_bundle.dart
-  /// in SDK code.
-  String? packageUriToServerPath(Uri packageUri) {
-    final defaultServerPath = '/packages/${packageUri.path}';
-    if (packageUri.isScheme('package')) {
-      if (!useDebuggerModuleNames) {
-        return defaultServerPath;
+  /// Computes the application URI (e.g., package: URI) for a given server path.
+  String? serverPathToAppUri(String serverPath);
+
+  /// Computes the resolved file URI for a server path.
+  Uri? serverPathToResolvedUri(String serverPath);
+
+  PackageConfig? get packageConfig;
+
+  bool get useDebuggerModuleNames;
+}
+
+class FrontendServerPathResolver implements PathResolver {
+  final _logger = Logger('FrontendServerPathResolver');
+  @override
+  final PackageConfig? packageConfig;
+  @override
+  final bool useDebuggerModuleNames;
+  final String? packageRoot;
+
+  FrontendServerPathResolver({
+    this.packageConfig,
+    this.useDebuggerModuleNames = false,
+    this.packageRoot,
+  });
+
+  @override
+  String? appUriToServerPath(String appUrl, {bool? useDebuggerModuleNames}) {
+    final useDebugger = useDebuggerModuleNames ?? this.useDebuggerModuleNames;
+    final appUri = Uri.parse(appUrl);
+    // Note: must match `urlForComponentUri` in javascript_bundle.dart in SDK.
+    if (appUri.isScheme('package')) {
+      // package:foo/bar.dart -> packages/foo/lib/bar.dart (useDebugger)
+      // package:foo/bar.dart -> packages/foo/bar.dart (!useDebugger)
+      final pathSegments = appUri.pathSegments;
+      if (pathSegments.isEmpty) {
+        throw FormatException('Invalid package URI with empty path: $appUrl');
       }
-      final resolvedUri = packageConfig.resolve(packageUri);
-      if (resolvedUri == null) {
-        _logger.severe('Cannot resolve package uri $packageUri');
-        return defaultServerPath;
-      }
-      final package = packageConfig.packageOf(resolvedUri);
-      if (package == null) {
-        _logger.severe('Cannot find package for package uri $packageUri');
-        return defaultServerPath;
-      }
-      final root = package.root;
-      final relativeUrl = resolvedUri.toString().replaceFirst('$root', '');
-      final relativeRoot = _getRelativeRoot(root);
-      final ret = relativeRoot == null
-          ? 'packages/$relativeUrl'
-          : 'packages/$relativeRoot/$relativeUrl';
-      return ret;
+      final buildRunnerPath = 'packages/${appUri.path}';
+      final path = useDebugger
+          ? WebPathTranslator.addLibSegment(buildRunnerPath)
+          : buildRunnerPath;
+      return path;
     }
-    _logger.severe('Expected package uri, but found $packageUri');
+    if (appUri.isScheme('org-dartlang-app')) {
+      // org-dartlang-app:///web/main.dart -> web/main.dart
+      // org-dartlang-app:///packages/foo/bar.dart -> packages/foo/lib/bar.dart (useDebugger)
+      // org-dartlang-app:///packages/foo/lib/bar.dart -> packages/foo/bar.dart (!useDebugger)
+      final segments = appUri.pathSegments;
+      if (segments.isEmpty) {
+        throw FormatException('Invalid org-dartlang-app URI: $appUrl');
+      }
+      final path = useDebugger
+          ? WebPathTranslator.addLibSegment(appUri.path.substring(1))
+          : WebPathTranslator.removeLibSegment(appUri.path.substring(1));
+      return path;
+    }
     return null;
   }
 
-  /// Compute resolved file URI for a server path.
-  ///
-  /// Frontend Server serves package URIs with 'lib' intact, representing their
-  /// on-disk structure. For example, 'package:foo/bar.dart' would be served at
-  /// 'packages/foo/lib/bar.dart'.
-  ///
-  /// Build runner serves package URIs from a flattened 'packages' directory.
-  /// For example, 'package:foo/bar.dart' would be served at
-  /// 'packages/foo/bar.dart'.
+  @override
+  String? serverPathToAppUri(String serverPath) {
+    // packages/foo/lib/bar.dart -> package:foo/bar.dart
+    // packages/foo/bar.dart -> package:foo/bar.dart
+    // web/main.dart -> web/main.dart
+    serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
+    if (!serverPath.startsWith('packages/')) return serverPath;
+    final pathWithoutLib = WebPathTranslator.removeLibSegment(serverPath);
+    return pathWithoutLib.replaceFirst('packages/', 'package:');
+  }
+
+  @override
   Uri? serverPathToResolvedUri(String serverPath) {
     serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
     final segments = serverPath.split('/');
     if (segments.first == 'packages') {
-      final packagePath = WebPathTranslator.translatePackagesPathToPackageUri(
-        serverPath,
-      );
-      return packageConfig.resolve(Uri.parse(packagePath));
+      final config = packageConfig;
+      if (config == null) {
+        _logger.severe('Cannot resolve packages without packageConfig');
+        return null;
+      }
+      final packagePath = serverPathToAppUri(serverPath);
+      if (packagePath == null) return null;
+      return config.resolve(Uri.parse(packagePath));
+    } else if (packageRoot != null) {
+      return Uri.file(p.join(packageRoot!, serverPath));
     }
-    _logger.severe('Expected "packages/" path, but found $serverPath');
+    _logger.severe(
+      'Cannot resolve path without packages/ prefix or packageRoot: $serverPath',
+    );
     return null;
   }
 }
 
-String? _getRelativeRoot(Uri root) =>
-    root.pathSegments.lastWhereOrNull((segment) => segment.isNotEmpty);
+class BuildRunnerPathResolver implements PathResolver {
+  final _logger = Logger('BuildRunnerPathResolver');
+  @override
+  final PackageConfig? packageConfig;
+  @override
+  final bool useDebuggerModuleNames;
+  final String? packageRoot;
+
+  BuildRunnerPathResolver({
+    this.packageConfig,
+    this.useDebuggerModuleNames = false,
+    this.packageRoot,
+  });
+
+  @override
+  String? appUriToServerPath(String appUrl, {bool? useDebuggerModuleNames}) {
+    final useDebugger = useDebuggerModuleNames ?? this.useDebuggerModuleNames;
+    final appUri = Uri.parse(appUrl);
+    // Note: must match `urlForComponentUri` in javascript_bundle.dart in SDK.
+    if (appUri.isScheme('package')) {
+      // package:foo/bar.dart -> packages/foo/bar.dart (useDebugger)
+      // package:foo/bar.dart -> /packages/foo/bar.dart (!useDebugger)
+      final pathSegments = appUri.pathSegments;
+      if (pathSegments.isEmpty) {
+        throw FormatException('Invalid package URI with empty path: $appUrl');
+      }
+      final path = 'packages/${appUri.path}';
+      if (!useDebugger && path.startsWith('packages/')) {
+        return '/$path';
+      }
+      return path;
+    }
+
+    if (appUri.isScheme('org-dartlang-app')) {
+      // org-dartlang-app:///web/main.dart -> main.dart
+      // org-dartlang-app:///packages/foo/bar.dart -> packages/foo/bar.dart
+      final segments = appUri.pathSegments;
+      if (segments.isEmpty) {
+        throw FormatException('Invalid org-dartlang-app URI: $appUrl');
+      }
+      final first = segments.first;
+      if (first == 'packages') {
+        if (segments.length < 3) {
+          throw FormatException('Invalid package path in app URI: $appUrl');
+        }
+        return segments.join('/');
+      }
+      return segments.skip(1).join('/');
+    }
+
+    return null;
+  }
+
+  @override
+  String? serverPathToAppUri(String serverPath) {
+    // packages/foo/bar.dart -> package:foo/bar.dart
+    // web/main.dart -> web/main.dart
+    serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
+    if (!serverPath.startsWith('packages/')) return serverPath;
+    return serverPath.replaceFirst('packages/', 'package:');
+  }
+
+  @override
+  Uri? serverPathToResolvedUri(String serverPath) {
+    serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
+    final segments = serverPath.split('/');
+    if (segments.first == 'packages') {
+      final config = packageConfig;
+      if (config == null) {
+        _logger.severe('Cannot resolve packages without packageConfig');
+        return null;
+      }
+      final packagePath = serverPathToAppUri(serverPath);
+      if (packagePath == null) return null;
+      return config.resolve(Uri.parse(packagePath));
+    } else if (packageRoot != null) {
+      return Uri.file(p.join(packageRoot!, serverPath));
+    }
+    _logger.severe(
+      'Cannot resolve path without packages/ prefix or packageRoot: $serverPath',
+    );
+    return null;
+  }
+}
+
+class FlutterPathResolver implements PathResolver {
+  final _logger = Logger('FlutterPathResolver');
+  @override
+  final PackageConfig? packageConfig;
+  @override
+  final bool useDebuggerModuleNames;
+  final String? packageRoot;
+
+  FlutterPathResolver({
+    this.packageConfig,
+    this.useDebuggerModuleNames = false,
+    this.packageRoot,
+  });
+
+  @override
+  String? appUriToServerPath(String appUrl, {bool? useDebuggerModuleNames}) {
+    final useDebugger = useDebuggerModuleNames ?? this.useDebuggerModuleNames;
+    final isFlutterPackage = appUrl.startsWith('package:');
+
+    if (isFlutterPackage) {
+      // package:foo/bar.dart -> packages/foo/bar.dart
+      final appUri = Uri.parse(appUrl);
+      final pathSegments = appUri.pathSegments;
+      if (pathSegments.isEmpty) {
+        throw FormatException('Invalid package URI with empty path: $appUrl');
+      }
+      final path = 'packages/${appUri.path}';
+      return path;
+    } else {
+      final appUri = Uri.parse(appUrl);
+      if (appUri.isScheme('org-dartlang-app')) {
+        // org-dartlang-app:///web/main.dart -> web/main.dart
+        // org-dartlang-app:///packages/foo/bar.dart -> packages/foo/lib/bar.dart (useDebugger)
+        // org-dartlang-app:///packages/foo/lib/bar.dart -> packages/foo/bar.dart (!useDebugger)
+        final path = useDebugger
+            ? WebPathTranslator.addLibSegment(appUri.path.substring(1))
+            : WebPathTranslator.removeLibSegment(appUri.path.substring(1));
+        return path;
+      }
+    }
+    return null;
+  }
+
+  @override
+  String? serverPathToAppUri(String serverPath) {
+    // packages/foo/lib/bar.dart -> package:foo/bar.dart
+    // packages/foo/bar.dart -> package:foo/bar.dart
+    // web/main.dart -> web/main.dart
+    serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
+    if (!serverPath.startsWith('packages/')) return serverPath;
+    final pathWithoutLib = WebPathTranslator.removeLibSegment(serverPath);
+    return pathWithoutLib.replaceFirst('packages/', 'package:');
+  }
+
+  @override
+  Uri? serverPathToResolvedUri(String serverPath) {
+    serverPath = stripLeadingSlashes(serverPath).replaceAll('\\', '/');
+    final segments = serverPath.split('/');
+    if (segments.first == 'packages') {
+      final config = packageConfig;
+      if (config == null) {
+        _logger.severe('Cannot resolve packages without packageConfig');
+        return null;
+      }
+      final packagePath = serverPathToAppUri(serverPath);
+      if (packagePath == null) return null;
+      return config.resolve(Uri.parse(packagePath));
+    } else if (packageRoot != null) {
+      return Uri.file(p.join(packageRoot!, serverPath));
+    }
+    _logger.severe(
+      'Cannot resolve path without packages/ prefix or packageRoot: $serverPath',
+    );
+    return null;
+  }
+}