+++ dwds/lib/src/readers/frontend_server_asset_reader.dart
@@ -6,7 +6,6 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:dwds/src/readers/asset_reader.dart';
-import 'package:dwds/src/utilities/web_path_translator.dart';
 import 'package:logging/logging.dart';
 import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
@@ -66,11 +65,12 @@ class FrontendServerAssetReader implements AssetReader {
       var strippedPath = _stripBasePath(serverPath);
       Uri? fileUri;
       if (strippedPath.startsWith('packages/')) {
-        final packagePath = WebPathTranslator.translatePackagesPathToPackageUri(
+        final packagePath = FrontendServerPathResolver().serverPathToAppUri(
           strippedPath,
-          layout: AppUriLayout.frontendServerOnly,
         );
-        fileUri = packageConfig.resolve(Uri.parse(packagePath));
+        if (packagePath != null) {
+          fileUri = packageConfig.resolve(Uri.parse(packagePath));
+        }
       } else {
         fileUri = p.toUri(p.join(_packageRoot, strippedPath));
       }