+++ dwds/test/integration/metadata_test.dart
@@ -7,6 +7,7 @@ library;
 
 import 'dart:convert';
 
+import 'package:dwds/src/debugging/metadata/loader.dart';
 import 'package:dwds/src/debugging/metadata/module_metadata.dart';
 import 'package:dwds/src/debugging/metadata/provider.dart';
 import 'package:test/test.dart';
@@ -40,7 +41,7 @@ void main() {
   test('can parse metadata with empty sources', () async {
     final provider = MetadataProvider(
       'foo.bootstrap.js',
-      FakeAssetReader(metadata: _emptySourceMetadata),
+      MergedMetadataLoader(FakeAssetReader(metadata: _emptySourceMetadata)),
     );
     expect(
       await provider.libraries,
@@ -51,7 +52,7 @@ void main() {
   test('throws on metadata with absolute import uris', () async {
     final provider = MetadataProvider(
       'foo.bootstrap.js',
-      FakeAssetReader(metadata: _fileUriMetadata),
+      MergedMetadataLoader(FakeAssetReader(metadata: _fileUriMetadata)),
     );
     await expectLater(
       provider.libraries,
@@ -64,7 +65,7 @@ void main() {
     () async {
       final provider = MetadataProvider(
         'foo.bootstrap.js',
-        FakeAssetReader(metadata: _emptySourceMetadata),
+        MergedMetadataLoader(FakeAssetReader(metadata: _emptySourceMetadata)),
       );
       final modulePath = 'foo/web/main.ddc.js';
       final moduleName = 'web/main';
@@ -210,7 +211,10 @@ void main() {
     final assetReader = FakeAssetReader(
       metadata: createMetadataContents(moduleToLibraries, libraryToParts),
     );
-    final provider = MetadataProvider('foo.bootstrap.js', assetReader);
+    final provider = MetadataProvider(
+      'foo.bootstrap.js',
+      MergedMetadataLoader(assetReader),
+    );
     await validateProvider(provider, moduleToLibraries, libraryToParts);
 
     const newModuleToLibraries = <String, List<String>>{