+++ dwds/test/utilities/web_path_translator_test.dart
@@ -2,128 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:dwds/src/loaders/asset_scheme.dart';
 import 'package:dwds/src/utilities/web_path_translator.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('WebPathTranslator', () {
-    group('translateAppUriToServerPath', () {
-      group('package: URIs', () {
-        test('frontendServerOnly layout adds lib/ segment', () {
-          // Checks: package:foo/bar.dart -> packages/foo/lib/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'package:foo/bar.dart',
-              layout: AppUriLayout.frontendServerOnly,
-            ),
-            'packages/foo/lib/bar.dart',
-          );
-        });
-
-        test('frontendServerOnly layout omits lib/ segment if useDebuggerModuleNames is false', () {
-          // Checks: package:foo/bar.dart -> packages/foo/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'package:foo/bar.dart',
-              layout: AppUriLayout.frontendServerOnly,
-              useDebuggerModuleNames: false,
-            ),
-            'packages/foo/bar.dart',
-          );
-        });
-
-        test('buildRunner layout omits lib/ segment', () {
-          // Checks: package:foo/bar.dart -> packages/foo/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'package:foo/bar.dart',
-              layout: AppUriLayout.buildRunner,
-            ),
-            'packages/foo/bar.dart',
-          );
-        });
-
-        test('flutter layout omits lib/ segment', () {
-          // Checks: package:foo/bar.dart -> packages/foo/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'package:foo/bar.dart',
-              layout: AppUriLayout.flutter,
-            ),
-            'packages/foo/bar.dart',
-          );
-        });
-      });
-
-      group('org-dartlang-app: URIs', () {
-        test('frontendServerOnly layout preserves path', () {
-          // Checks: org-dartlang-app:///web/main.dart -> web/main.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///web/main.dart',
-              layout: AppUriLayout.frontendServerOnly,
-            ),
-            'web/main.dart',
-          );
-        });
-
-        test('frontendServerOnly layout adds lib/ to packages path', () {
-          // Checks: org-dartlang-app:///packages/foo/bar.dart -> packages/foo/lib/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///packages/foo/bar.dart',
-              layout: AppUriLayout.frontendServerOnly,
-            ),
-            'packages/foo/lib/bar.dart',
-          );
-        });
-
-        test('buildRunner layout strips entrypoint prefix', () {
-          // Checks: org-dartlang-app:///web/main.dart -> main.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///web/main.dart',
-              layout: AppUriLayout.buildRunner,
-            ),
-            'main.dart',
-          );
-        });
-
-        test('buildRunner layout preserves packages path', () {
-          // Checks: org-dartlang-app:///packages/foo/bar.dart -> packages/foo/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///packages/foo/bar.dart',
-              layout: AppUriLayout.buildRunner,
-            ),
-            'packages/foo/bar.dart',
-          );
-        });
-
-        test('flutter layout behaves like buildRunner for entrypoints', () {
-          // Checks: org-dartlang-app:///web/main.dart -> main.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///web/main.dart',
-              layout: AppUriLayout.flutter,
-            ),
-            'main.dart',
-          );
-        });
-
-        test('flutter layout behaves like buildRunner for packages path', () {
-          // Checks: org-dartlang-app:///packages/foo/bar.dart -> packages/foo/bar.dart
-          expect(
-            WebPathTranslator.translateAppUriToServerPath(
-              'org-dartlang-app:///packages/foo/bar.dart',
-              layout: AppUriLayout.flutter,
-            ),
-            'packages/foo/bar.dart',
-          );
-        });
-      });
-    });
-
     group('addLibSegment', () {
       test('adds lib/ to packages paths', () {
         // Checks: packages/foo/bar.dart -> packages/foo/lib/bar.dart
@@ -171,52 +55,6 @@ void main() {
         // Checks: web/main.dart -> web/main.dart
         expect(
           WebPathTranslator.removeLibSegment('web/main.dart'),
-          'web/main.dart',
-        );
-      });
-    });
-
-    group('translatePackagesPathToPackageUri', () {
-      test('translates packages/ paths with lib/ (frontendServerOnly)', () {
-        // Checks: packages/foo/lib/bar.dart -> package:foo/bar.dart
-        expect(
-          WebPathTranslator.translatePackagesPathToPackageUri(
-            'packages/foo/lib/bar.dart',
-            layout: AppUriLayout.frontendServerOnly,
-          ),
-          'package:foo/bar.dart',
-        );
-      });
-
-      test('translates packages/ paths without lib/ (buildRunner)', () {
-        // Checks: packages/foo/bar.dart -> package:foo/bar.dart
-        expect(
-          WebPathTranslator.translatePackagesPathToPackageUri(
-            'packages/foo/bar.dart',
-            layout: AppUriLayout.buildRunner,
-          ),
-          'package:foo/bar.dart',
-        );
-      });
-
-      test('preserves lib/ if present in buildRunner layout', () {
-        // Checks: packages/foo/lib/bar.dart -> package:foo/lib/bar.dart
-        // Note: we usually never want 'lib' to be retained in a package path.
-        // However, this tests that a file in a nested 'lib/' directory (e.g.,
-        // 'lib/lib/bar.dart') is reconstructed properly.
-        expect(
-          WebPathTranslator.translatePackagesPathToPackageUri(
-            'packages/foo/lib/bar.dart',
-            layout: AppUriLayout.buildRunner,
-          ),
-          'package:foo/lib/bar.dart',
-        );
-      });
-
-      test('is no-op for non-packages paths', () {
-        // Checks: web/main.dart -> web/main.dart
-        expect(
-          WebPathTranslator.translatePackagesPathToPackageUri('web/main.dart'),
           'web/main.dart',
         );
       });
@@ -287,8 +125,8 @@ void main() {
           expect(
             WebPathTranslator.translateModuleExtension(
               'main.dart.lib',
-              from: AppUriLayout.frontendServerOnly,
-              to: AppUriLayout.buildRunner,
+              from: FrontendServerAssetScheme(),
+              to: BuildRunnerAssetScheme(),
             ),
             'main.ddc',
           );
@@ -296,8 +134,8 @@ void main() {
           expect(
             WebPathTranslator.translateModuleExtension(
               'main.dart.lib.js',
-              from: AppUriLayout.frontendServerOnly,
-              to: AppUriLayout.buildRunner,
+              from: FrontendServerAssetScheme(),
+              to: BuildRunnerAssetScheme(),
             ),
             'main.ddc.js',
           );
@@ -311,8 +149,8 @@ void main() {
           expect(
             WebPathTranslator.translateModuleExtension(
               'main.ddc',
-              from: AppUriLayout.buildRunner,
-              to: AppUriLayout.frontendServerOnly,
+              from: BuildRunnerAssetScheme(),
+              to: FrontendServerAssetScheme(),
             ),
             'main.dart.lib',
           );
@@ -320,8 +158,8 @@ void main() {
           expect(
             WebPathTranslator.translateModuleExtension(
               'main.ddc.js',
-              from: AppUriLayout.buildRunner,
-              to: AppUriLayout.frontendServerOnly,
+              from: BuildRunnerAssetScheme(),
+              to: FrontendServerAssetScheme(),
             ),
             'main.dart.lib.js',
           );
@@ -333,8 +171,8 @@ void main() {
         expect(
           WebPathTranslator.translateModuleExtension(
             'main.dart.lib',
-            from: AppUriLayout.frontendServerOnly,
-            to: AppUriLayout.frontendServerOnly,
+            from: FrontendServerAssetScheme(),
+            to: FrontendServerAssetScheme(),
           ),
           'main.dart.lib',
         );